backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
better-exceptions==0.3.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.13.1